Move LaunchPredictor to state_estimation
